requirements.txt
aiofiles==23.2.1 ; python_version >= "3.8" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.8" and python_version < "4.0"
ansicon==1.89.0 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
anyio==4.3.0 ; python_version >= "3.8" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.8" and python_version < "4.0"
blessed==1.20.0 ; python_version >= "3.8" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.8" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
h2o-authn==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
h2o-wave==1.1.1 ; python_version >= "3.8" and python_version < "4.0"
h2ogpte==1.4.10 ; python_version >= "3.8" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.8" and python_version < "4.0"
httpx==0.26.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.6 ; python_version >= "3.8" and python_version < "4.0"
importlib-metadata==7.1.0 ; python_version >= "3.8" and python_version < "3.10"
inquirer==3.1.4 ; python_version >= "3.8" and python_version < "4.0"
jinxed==1.2.1 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
lxml==4.9.3 ; python_version >= "3.8" and python_version < "4.0"
markdown==3.6 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.9"
numpy==1.26.4 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.8" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.8" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic-core==2.14.5 ; python_version >= "3.8" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.8" and python_version < "4.0"
pydantic==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
pydantic[dotenv]==2.5.2 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.8" and python_version < "4.0"
python-editor==1.0.4 ; python_version >= "3.8" and python_version < "4.0"
python-pptx==0.6.23 ; python_version >= "3.8" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.8" and python_version < "4.0"
readchar==4.0.6 ; python_version >= "3.8" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==69.2.0 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.8" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.8" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.8" and python_version < "4.0"
uvicorn==0.29.0 ; python_version >= "3.8" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.8" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.8" and python_version < "4.0"
xlsxwriter==3.2.0 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.18.1 ; python_version >= "3.8" and python_version < "3.10"
tqdm
toml
Levenshtein
Spire.PDF
Spire.Presentation
pyngrok
